Fixes for 32-bit systems
* added cast_checked() macro to runtime-assert an integer narrowing
* added HAVE_STRUCT_INT128 and fixed wherever __int128_t was being
  treated as a an integer
* replaced occurences of %lu with %zu for size_t parameters
* replaced pointer-to-uint64_t casts with casts to uintptr_t
* added uintptr2_t type for aliasing atomic (pointer,int) type

+++ src/libponyc/ast/parserapi.h
@@ -413,7 +413,7 @@ bool parse(ast_t* package, source_t* source, rule_t start,
  *      SET_FLAG(FOO_FLAG);
  */
 #define SET_FLAG(f) \
-  ast_setdata(state.ast, (void*)(f | (uint64_t)ast_data(state.ast)))
+  ast_setdata(state.ast, (void*)(f | (uintptr_t)ast_data(state.ast)))
 
 
 /** Set the data field flags to use for the next token found in the source.
